msgpack: move pack and unpack functions to msgpack_export_test.go
